+++ onAirStudy/src/main/resources/mapper/invitation/invitation-mapper.xml
@@ -4,16 +4,35 @@
   "http://mybatis.org/dtd/mybatis-3-mapper.dtd">
 <mapper namespace="invitation">
 
-<select id="selectInvitationList" resultType="invitation">
+<select id="selectInvitationList" resultMap="invitationMap">
 
 	select  
-		*
+		I.*,
+		(select     sr_title
+    	from    sr_info
+    	where sr_no =I.sr_no) s_title,
+    	(select     sr_goal
+    	from    sr_info
+    	where sr_no =I.sr_no) s_goal
+    	
 	from
-		invitation I join sr_info S
-	using
-		(sr_no)
+		invitation I
+	
 	order by
 		no desc
 
 </select>
+
+<resultMap type="map" id="invitationMap">
+  		<id column="no" property="no"/>
+  		<result column="hosting_id" property="hostingId"/>
+  		<result column="invited_id" property="invitedId"/>
+  		<result column="sr_no" property="srNo"/>
+  		<result column="invitation_date" property="invitationDate"/>
+  		<result column="invitation_yn" property="invitationYN"/>
+  		<result column="sr_ended_yn" property="srEndedYN"/>
+  		<result column="s_title" property="sTitle"/>
+  		<result column="s_goal" property="sGoal"/>
+  		
+  	</resultMap>
 </mapper>
