+++ onAirStudy/src/main/webapp/WEB-INF/jspwork/org/apache/jsp/WEB_002dINF/views/invitation/invitationListForm_jsp.java
@@ -1,0 +1,380 @@
+/*
+ * Generated by the Jasper component of Apache Tomcat
+ * Version: Apache Tomcat/8.5.57
+ * Generated at: 2020-10-18 15:25:13 UTC
+ * Note: The last modified time of this file was set to
+ *       the last modified time of the source file after
+ *       generation to assist with modification tracking.
+ */
+package org.apache.jsp.WEB_002dINF.views.invitation;
+
+import javax.servlet.*;
+import javax.servlet.http.*;
+import javax.servlet.jsp.*;
+
+public final class invitationListForm_jsp extends org.apache.jasper.runtime.HttpJspBase
+    implements org.apache.jasper.runtime.JspSourceDependent,
+                 org.apache.jasper.runtime.JspSourceImports {
+
+  private static final javax.servlet.jsp.JspFactory _jspxFactory =
+          javax.servlet.jsp.JspFactory.getDefaultFactory();
+
+  private static java.util.Map<java.lang.String,java.lang.Long> _jspx_dependants;
+
+  static {
+    _jspx_dependants = new java.util.HashMap<java.lang.String,java.lang.Long>(4);
+    _jspx_dependants.put("/WEB-INF/lib/taglibs-standard-impl-1.2.5.jar", Long.valueOf(1599824644000L));
+    _jspx_dependants.put("jar:file:/C:/GitRepo/final/onAirStudy/onAirStudy/src/main/webapp/WEB-INF/lib/taglibs-standard-impl-1.2.5.jar!/META-INF/fn.tld", Long.valueOf(1425946270000L));
+    _jspx_dependants.put("jar:file:/C:/GitRepo/final/onAirStudy/onAirStudy/src/main/webapp/WEB-INF/lib/taglibs-standard-impl-1.2.5.jar!/META-INF/fmt.tld", Long.valueOf(1425946270000L));
+    _jspx_dependants.put("jar:file:/C:/GitRepo/final/onAirStudy/onAirStudy/src/main/webapp/WEB-INF/lib/taglibs-standard-impl-1.2.5.jar!/META-INF/c.tld", Long.valueOf(1425946270000L));
+  }
+
+  private static final java.util.Set<java.lang.String> _jspx_imports_packages;
+
+  private static final java.util.Set<java.lang.String> _jspx_imports_classes;
+
+  static {
+    _jspx_imports_packages = new java.util.HashSet<>();
+    _jspx_imports_packages.add("javax.servlet");
+    _jspx_imports_packages.add("javax.servlet.http");
+    _jspx_imports_packages.add("javax.servlet.jsp");
+    _jspx_imports_classes = null;
+  }
+
+  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems;
+  private org.apache.jasper.runtime.TagHandlerPool _005fjspx_005ftagPool_005fc_005fif_0026_005ftest;
+
+  private volatile javax.el.ExpressionFactory _el_expressionfactory;
+  private volatile org.apache.tomcat.InstanceManager _jsp_instancemanager;
+
+  public java.util.Map<java.lang.String,java.lang.Long> getDependants() {
+    return _jspx_dependants;
+  }
+
+  public java.util.Set<java.lang.String> getPackageImports() {
+    return _jspx_imports_packages;
+  }
+
+  public java.util.Set<java.lang.String> getClassImports() {
+    return _jspx_imports_classes;
+  }
+
+  public javax.el.ExpressionFactory _jsp_getExpressionFactory() {
+    if (_el_expressionfactory == null) {
+      synchronized (this) {
+        if (_el_expressionfactory == null) {
+          _el_expressionfactory = _jspxFactory.getJspApplicationContext(getServletConfig().getServletContext()).getExpressionFactory();
+        }
+      }
+    }
+    return _el_expressionfactory;
+  }
+
+  public org.apache.tomcat.InstanceManager _jsp_getInstanceManager() {
+    if (_jsp_instancemanager == null) {
+      synchronized (this) {
+        if (_jsp_instancemanager == null) {
+          _jsp_instancemanager = org.apache.jasper.runtime.InstanceManagerFactory.getInstanceManager(getServletConfig());
+        }
+      }
+    }
+    return _jsp_instancemanager;
+  }
+
+  public void _jspInit() {
+    _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
+    _005fjspx_005ftagPool_005fc_005fif_0026_005ftest = org.apache.jasper.runtime.TagHandlerPool.getTagHandlerPool(getServletConfig());
+  }
+
+  public void _jspDestroy() {
+    _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.release();
+    _005fjspx_005ftagPool_005fc_005fif_0026_005ftest.release();
+  }
+
+  public void _jspService(final javax.servlet.http.HttpServletRequest request, final javax.servlet.http.HttpServletResponse response)
+      throws java.io.IOException, javax.servlet.ServletException {
+
+    final java.lang.String _jspx_method = request.getMethod();
+    if (!"GET".equals(_jspx_method) && !"POST".equals(_jspx_method) && !"HEAD".equals(_jspx_method) && !javax.servlet.DispatcherType.ERROR.equals(request.getDispatcherType())) {
+      response.sendError(HttpServletResponse.SC_METHOD_NOT_ALLOWED, "JSPs only permit GET, POST or HEAD. Jasper also permits OPTIONS");
+      return;
+    }
+
+    final javax.servlet.jsp.PageContext pageContext;
+    javax.servlet.http.HttpSession session = null;
+    final javax.servlet.ServletContext application;
+    final javax.servlet.ServletConfig config;
+    javax.servlet.jsp.JspWriter out = null;
+    final java.lang.Object page = this;
+    javax.servlet.jsp.JspWriter _jspx_out = null;
+    javax.servlet.jsp.PageContext _jspx_page_context = null;
+
+
+    try {
+      response.setContentType("text/html; charset=UTF-8");
+      pageContext = _jspxFactory.getPageContext(this, request, response,
+      			null, true, 8192, true);
+      _jspx_page_context = pageContext;
+      application = pageContext.getServletContext();
+      config = pageContext.getServletConfig();
+      session = pageContext.getSession();
+      out = pageContext.getOut();
+      _jspx_out = out;
+
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("<!DOCTYPE html>\r\n");
+      out.write("<html>\r\n");
+      out.write("<head>\r\n");
+      out.write("\r\n");
+      out.write("  <meta charset=\"UTF-8\">\r\n");
+      out.write("  <meta name=\"viewport\" content=\"width=device-width, initial-scale=1.0\">\r\n");
+      out.write("  <title>초대내역- Invitation list</title>\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("  <link rel=\"stylesheet\" href=\"https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/css/bootstrap.min.css\"\r\n");
+      out.write("    integrity=\"sha384-JcKb8q3iqJ61gNV9KGb8thSsNjpSL0n8PARn9HuZOnIxN0hoP+VmmDGMN5t9UJ0Z\" crossorigin=\"anonymous\">\r\n");
+      out.write("  <script src=\"https://stackpath.bootstrapcdn.com/bootstrap/4.5.2/js/bootstrap.min.js\"\r\n");
+      out.write("    integrity=\"sha384-B4gt1jrGC7Jh4AgTPSdUtOBvfO8shuf57BaghqFfPlYxofvL8/KUEfYiJOMMV+rV\"\r\n");
+      out.write("    crossorigin=\"anonymous\"></script>\r\n");
+      out.write("  <script src=\"https://code.jquery.com/jquery-3.5.1.slim.min.js\"\r\n");
+      out.write("    integrity=\"sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj\"\r\n");
+      out.write("    crossorigin=\"anonymous\"></script>\r\n");
+      out.write("  <script src=\"https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js\"\r\n");
+      out.write("    integrity=\"sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN\"\r\n");
+      out.write("    crossorigin=\"anonymous\"></script>\r\n");
+      out.write("\r\n");
+      out.write(" <link rel=\"stylesheet\" href=\"");
+      out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${pageContext.request.contextPath }", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+      out.write("/resources/css/invitation.css\">\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("</head>\r\n");
+      out.write("\r\n");
+      out.write("<body>\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("  <div class=\"invitation\" style=\"margin-bottom: 0;\">\r\n");
+      out.write("\r\n");
+      out.write("   \r\n");
+      out.write("      \r\n");
+      out.write("      <div class=\"col-lg\" style=\"background-color:#FBF7FD; padding: 20px;\">\r\n");
+      out.write("        <h2 class=\"text-gray\">Invitation List</h2>\r\n");
+      out.write("        <hr>\r\n");
+      out.write("        <div class=\"container\">\r\n");
+      out.write("          <div class=\"col-sm\" style=\"background-color: #F9F1ED; border-radius: 25px; border: 2px solid #AD8686; padding: 20px;\">\r\n");
+      out.write("            <h3 >나를 초대한 스터디 그룹</h3>\r\n");
+      out.write("\r\n");
+      out.write("            <table class=\"table\">\r\n");
+      out.write("              <thead class=\"thead-rounded\">\r\n");
+      out.write("                <tr>\r\n");
+      out.write("                  <th>No.</th>                  \r\n");
+      out.write("                  <th>그룹 이름</th>\r\n");
+      out.write("                  <th>그룹 소개</th>\r\n");
+      out.write("                  <th>초대자</th>\r\n");
+      out.write("                  <th></th>\r\n");
+      out.write("                </tr>\r\n");
+      out.write("              </thead>\r\n");
+      out.write("              <tbody>\r\n");
+      out.write("                 ");
+      if (_jspx_meth_c_005fforEach_005f0(_jspx_page_context))
+        return;
+      out.write(" \r\n");
+      out.write("           \r\n");
+      out.write("              </tbody>\r\n");
+      out.write("            </table>\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("          </div>\r\n");
+      out.write("\r\n");
+      out.write("        </div>\r\n");
+      out.write("      </div>\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("  \r\n");
+      out.write("  <!-- The Modal -->\r\n");
+      out.write("  <div class=\"modal fade\" id=\"myModal_ok\">\r\n");
+      out.write("    <div class=\"modal-dialog\">\r\n");
+      out.write("      <div class=\"modal-content\">\r\n");
+      out.write("\r\n");
+      out.write("        <!-- Modal Header -->\r\n");
+      out.write("        <div class=\"modal-header\" style=\"color: #E2A182;\">\r\n");
+      out.write("          <h4 class=\"modal-title\">초대 수락</h4>\r\n");
+      out.write("          <button type=\"button\" class=\"close\" data-dismiss=\"modal\">×</button>\r\n");
+      out.write("        </div>\r\n");
+      out.write("\r\n");
+      out.write("        <!-- Modal body -->\r\n");
+      out.write("        <div class=\"modal-body\">\r\n");
+      out.write("          초대를 수락 하시겠습니까?\r\n");
+      out.write("        </div>\r\n");
+      out.write("\r\n");
+      out.write("        <!-- Modal footer -->\r\n");
+      out.write("        <div class=\"modal-footer\">\r\n");
+      out.write("          <button type=\"submit\" class=\"btn btn-outline\">확인</button>\r\n");
+      out.write("          <button type=\"button\" class=\"btn btn-danger\" data-dismiss=\"modal\">취소</button>\r\n");
+      out.write("        </div>\r\n");
+      out.write("\r\n");
+      out.write("      </div>\r\n");
+      out.write("    </div>\r\n");
+      out.write("  </div>\r\n");
+      out.write("\r\n");
+      out.write("  <!-- The Modal -->\r\n");
+      out.write("  <div class=\"modal fade\" id=\"myModal_no\">\r\n");
+      out.write("    <div class=\"modal-dialog\">\r\n");
+      out.write("      <div class=\"modal-content\">\r\n");
+      out.write("\r\n");
+      out.write("        <!-- Modal Header -->\r\n");
+      out.write("        <div class=\"modal-header\">\r\n");
+      out.write("          <h4 class=\"modal-title\">초대 거절</h4>\r\n");
+      out.write("          <button type=\"button\" class=\"close\" data-dismiss=\"modal\">×</button>\r\n");
+      out.write("        </div>\r\n");
+      out.write("\r\n");
+      out.write("        <!-- Modal body -->\r\n");
+      out.write("        <div class=\"modal-body\">\r\n");
+      out.write("          초대를 거절 하시겠습니까?\r\n");
+      out.write("        </div>\r\n");
+      out.write("\r\n");
+      out.write("        <!-- Modal footer -->\r\n");
+      out.write("        <div class=\"modal-footer\">\r\n");
+      out.write("          <button type=\"submit\" class=\"btn btn-outline\">확인</button>\r\n");
+      out.write("          <button type=\"button\" class=\"btn btn-danger\" data-dismiss=\"modal\">취소</button>\r\n");
+      out.write("        </div>\r\n");
+      out.write("\r\n");
+      out.write("      </div>\r\n");
+      out.write("    </div>\r\n");
+      out.write("  </div>\r\n");
+      out.write("<script>\r\n");
+      out.write("function invitation_no(no){\r\n");
+      out.write("\t$(\"#myModal_no\").moda().find(\"[name=no]\").val(no);\r\n");
+      out.write("}\r\n");
+      out.write("</script>\r\n");
+      out.write("\r\n");
+      out.write("\r\n");
+      out.write("</body>\r\n");
+      out.write("\r\n");
+      out.write("</html>");
+    } catch (java.lang.Throwable t) {
+      if (!(t instanceof javax.servlet.jsp.SkipPageException)){
+        out = _jspx_out;
+        if (out != null && out.getBufferSize() != 0)
+          try {
+            if (response.isCommitted()) {
+              out.flush();
+            } else {
+              out.clearBuffer();
+            }
+          } catch (java.io.IOException e) {}
+        if (_jspx_page_context != null) _jspx_page_context.handlePageException(t);
+        else throw new ServletException(t);
+      }
+    } finally {
+      _jspxFactory.releasePageContext(_jspx_page_context);
+    }
+  }
+
+  private boolean _jspx_meth_c_005fforEach_005f0(javax.servlet.jsp.PageContext _jspx_page_context)
+          throws java.lang.Throwable {
+    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
+    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
+    //  c:forEach
+    org.apache.taglibs.standard.tag.rt.core.ForEachTag _jspx_th_c_005fforEach_005f0 = (org.apache.taglibs.standard.tag.rt.core.ForEachTag) _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.get(org.apache.taglibs.standard.tag.rt.core.ForEachTag.class);
+    boolean _jspx_th_c_005fforEach_005f0_reused = false;
+    try {
+      _jspx_th_c_005fforEach_005f0.setPageContext(_jspx_page_context);
+      _jspx_th_c_005fforEach_005f0.setParent(null);
+      // /WEB-INF/views/invitation/invitationListForm.jsp(57,17) name = items type = javax.el.ValueExpression reqTime = true required = false fragment = false deferredValue = true expectedTypeName = java.lang.Object deferredMethod = false methodSignature = null
+      _jspx_th_c_005fforEach_005f0.setItems(new org.apache.jasper.el.JspValueExpression("/WEB-INF/views/invitation/invitationListForm.jsp(57,17) '${ inList }'",_jsp_getExpressionFactory().createValueExpression(_jspx_page_context.getELContext(),"${ inList }",java.lang.Object.class)).getValue(_jspx_page_context.getELContext()));
+      // /WEB-INF/views/invitation/invitationListForm.jsp(57,17) name = var type = java.lang.String reqTime = false required = false fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
+      _jspx_th_c_005fforEach_005f0.setVar("invitation");
+      int[] _jspx_push_body_count_c_005fforEach_005f0 = new int[] { 0 };
+      try {
+        int _jspx_eval_c_005fforEach_005f0 = _jspx_th_c_005fforEach_005f0.doStartTag();
+        if (_jspx_eval_c_005fforEach_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
+          do {
+            out.write(" \r\n");
+            out.write("                ");
+            if (_jspx_meth_c_005fif_005f0(_jspx_th_c_005fforEach_005f0, _jspx_page_context, _jspx_push_body_count_c_005fforEach_005f0))
+              return true;
+            out.write("                \r\n");
+            out.write("                ");
+            int evalDoAfterBody = _jspx_th_c_005fforEach_005f0.doAfterBody();
+            if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
+              break;
+          } while (true);
+        }
+        if (_jspx_th_c_005fforEach_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
+          return true;
+        }
+      } catch (java.lang.Throwable _jspx_exception) {
+        while (_jspx_push_body_count_c_005fforEach_005f0[0]-- > 0)
+          out = _jspx_page_context.popBody();
+        _jspx_th_c_005fforEach_005f0.doCatch(_jspx_exception);
+      } finally {
+        _jspx_th_c_005fforEach_005f0.doFinally();
+      }
+      _005fjspx_005ftagPool_005fc_005fforEach_0026_005fvar_005fitems.reuse(_jspx_th_c_005fforEach_005f0);
+      _jspx_th_c_005fforEach_005f0_reused = true;
+    } finally {
+      org.apache.jasper.runtime.JspRuntimeLibrary.releaseTag(_jspx_th_c_005fforEach_005f0, _jsp_getInstanceManager(), _jspx_th_c_005fforEach_005f0_reused);
+    }
+    return false;
+  }
+
+  private boolean _jspx_meth_c_005fif_005f0(javax.servlet.jsp.tagext.JspTag _jspx_th_c_005fforEach_005f0, javax.servlet.jsp.PageContext _jspx_page_context, int[] _jspx_push_body_count_c_005fforEach_005f0)
+          throws java.lang.Throwable {
+    javax.servlet.jsp.PageContext pageContext = _jspx_page_context;
+    javax.servlet.jsp.JspWriter out = _jspx_page_context.getOut();
+    //  c:if
+    org.apache.taglibs.standard.tag.rt.core.IfTag _jspx_th_c_005fif_005f0 = (org.apache.taglibs.standard.tag.rt.core.IfTag) _005fjspx_005ftagPool_005fc_005fif_0026_005ftest.get(org.apache.taglibs.standard.tag.rt.core.IfTag.class);
+    boolean _jspx_th_c_005fif_005f0_reused = false;
+    try {
+      _jspx_th_c_005fif_005f0.setPageContext(_jspx_page_context);
+      _jspx_th_c_005fif_005f0.setParent((javax.servlet.jsp.tagext.Tag) _jspx_th_c_005fforEach_005f0);
+      // /WEB-INF/views/invitation/invitationListForm.jsp(58,16) name = test type = boolean reqTime = true required = true fragment = false deferredValue = false expectedTypeName = null deferredMethod = false methodSignature = null
+      _jspx_th_c_005fif_005f0.setTest(((java.lang.Boolean) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${invitation.invitedId eq sessionID }", boolean.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null)).booleanValue());
+      int _jspx_eval_c_005fif_005f0 = _jspx_th_c_005fif_005f0.doStartTag();
+      if (_jspx_eval_c_005fif_005f0 != javax.servlet.jsp.tagext.Tag.SKIP_BODY) {
+        do {
+          out.write("\r\n");
+          out.write("         \t        <tr>\r\n");
+          out.write("                   \t<td>");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${invitation.no}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("</td>                    \r\n");
+          out.write("                    <td>");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${invitation.sTitle}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("</td>\r\n");
+          out.write("                    <td>");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${invitation.sGoal}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("</td> \r\n");
+          out.write("                    <td>");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${invitation.hostingId}", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("</td> \r\n");
+          out.write("\r\n");
+          out.write("                    <td>\r\n");
+          out.write("                      <button type=\"button\" class=\"btn btn-outline-primary\" onclick=\"invitation_ok('");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ invitation.no }", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("')\">수락 </button>\r\n");
+          out.write("                      <button type=\"button\" class=\"btn btn-primary\" onclick=\"invitation_no('");
+          out.write((java.lang.String) org.apache.jasper.runtime.PageContextImpl.proprietaryEvaluate("${ invitation.no }", java.lang.String.class, (javax.servlet.jsp.PageContext)_jspx_page_context, null));
+          out.write("')\"> 거절 </button>\r\n");
+          out.write("                    </td>\r\n");
+          out.write("                  </tr> \r\n");
+          out.write("                  ");
+          int evalDoAfterBody = _jspx_th_c_005fif_005f0.doAfterBody();
+          if (evalDoAfterBody != javax.servlet.jsp.tagext.BodyTag.EVAL_BODY_AGAIN)
+            break;
+        } while (true);
+      }
+      if (_jspx_th_c_005fif_005f0.doEndTag() == javax.servlet.jsp.tagext.Tag.SKIP_PAGE) {
+        return true;
+      }
+      _005fjspx_005ftagPool_005fc_005fif_0026_005ftest.reuse(_jspx_th_c_005fif_005f0);
+      _jspx_th_c_005fif_005f0_reused = true;
+    } finally {
+      org.apache.jasper.runtime.JspRuntimeLibrary.releaseTag(_jspx_th_c_005fif_005f0, _jsp_getInstanceManager(), _jspx_th_c_005fif_005f0_reused);
+    }
+    return false;
+  }
+}
